Allow building/usage with sdk-7.3
For now, this only applies to bmdcapture and bmdplay;
genlock needs to set timing offset options which is different.
As genlock is not my primary focus, it hasn't been changed yet.

All behavior should be unchanged for recent SDKs, the code
may only behave different for sdk-7.3 where it did not work
at all previously. It has been tested on 7.3 though, so it should
be fine.