Permit really high memory limit (>4GB)
Reasoning: In some circumstances, the queue starts
growing larger and larger over time (several hours).
Although a memory limit certainly makes sense to avoid
crashing the entire system, it is useful for some cases
to be able to fully use the available RAM so that the
unwanted capture termination will happen as late as
possible.